Export remote snapshotter label handler
(cherry picked from commit dbf384a5a857b635e0ea8b160645be81e60b3858)
